Enum pty_proc::shell::device::control::operate::key::Key

pub enum Key {
    Char(u64),
    Str(In),
}

Variants
Char(u64)

Unicode characters.

Str(In)

Unicode strings.

Methods

impl Key

pub fn from_utf8(buf: [c_uchar; 4]) -> Self

The constructor method from_utf8 returns a UTF-8 parsed Key.

pub fn is_utf8(&self) -> Option<char>

pub fn is_up(&self) -> bool

The accessor method is_up returns an Option for the Up Key.

pub fn is_down(&self) -> bool

The accessor method is_down returns an Option for the down Key.

pub fn is_right(&self) -> bool

The accessor method is_right returns an Option for the right Key.

pub fn is_left(&self) -> bool

The accessor method is_left returns an Option for the left Key.

pub fn is_delete(&self) -> bool

The accessor method is_delete returns an Option for the delete Key.

pub fn is_backspace(&self) -> bool

The accessor method is_backspace returns an Option for the backspace Key.

pub fn is_escape(&self) -> bool

The accessor method is_escape returns an Option for the delete Key.

pub fn is_end(&self) -> bool

The accessor method is_end returns an Option for the delete Key.

pub fn is_home(&self) -> bool

The accessor method is_home returns an Option for the delete Key.

pub fn is_pageup(&self) -> bool

The accessor method is_pageup returns an Option for the Page Up Key.

pub fn is_pagedown(&self) -> bool

The accessor method is_pagedown returns an Option for the Page Down Key.

pub fn is_c0(&self) -> bool

pub fn is_null(&self) -> bool

pub fn is_start_heading(&self) -> bool

pub fn is_start_text(&self) -> bool

pub fn is_end_text(&self) -> bool

pub fn is_end_transmission(&self) -> bool

pub fn is_enquiry(&self) -> bool

pub fn is_acknowledge(&self) -> bool

pub fn is_bell(&self) -> bool

pub fn is_horizontal_tabulation(&self) -> bool

pub fn is_enter(&self) -> bool

The accessor method is_enter returns an Option for the Enter Key.

pub fn is_line_feed(&self) -> bool

pub fn is_vertical_tabulation(&self) -> bool

pub fn is_form_feed(&self) -> bool

pub fn is_carriage_return(&self) -> bool

pub fn is_shift_out(&self) -> bool

pub fn is_shift_in(&self) -> bool

pub fn is_data_link_escape(&self) -> bool

pub fn is_device_control_one(&self) -> bool

pub fn is_device_control_two(&self) -> bool

pub fn is_device_control_three(&self) -> bool

pub fn is_device_control_four(&self) -> bool

pub fn is_negative_acknowledge(&self) -> bool

pub fn is_synchronous_idle(&self) -> bool

pub fn is_end_transmission_block(&self) -> bool

pub fn is_cancel(&self) -> bool

pub fn is_end_of_medium(&self) -> bool

pub fn is_substitute(&self) -> bool

pub fn is_file_separator(&self) -> bool

pub fn is_group_separator(&self) -> bool

pub fn is_record_separator(&self) -> bool

pub fn is_unit_separator(&self) -> bool

pub fn is_space(&self) -> bool

pub fn is_c1(&self) -> bool

pub fn as_input(&self) -> (In, size_t)