Struct pty_proc::shell::display::cursor::Cursor

pub struct Cursor<T> { /* fields omitted */ }

Methods

impl<T> Cursor<T>

pub fn new(inner: T) -> Cursor<T>

Create a new cursor wrapping the provided underlying I/O object.

pub fn into_inner(self) -> T

Consume this cursor, returning the underlying value.

Important traits for &'a mut R
impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, W> Write for &'a mut W where
W: Write + ?Sized, impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn get_ref(&self) -> &T

Get a reference to the underlying value in this cursor.

Important traits for &'a mut R
impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, W> Write for &'a mut W where
W: Write + ?Sized, impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn get_mut(&mut self) -> &mut T

Get a mutable reference to the underlying value in this cursor.

Care should be taken to avoid modifying the internal I/O state of the underlying value as it may corrupt this cursor's position.

pub fn position(&self) -> usize

Returns the current value of this cursor

pub fn set_position(&mut self, pos: usize)

Sets the value of this cursor

impl Cursor<Vec<Character>>

pub fn write_with_color(&mut self, glyph: char, ope: Character) -> Result<usize>