Struct pty_proc::shell::display::Display

pub struct Display { /* fields omitted */ }

Methods

impl Display

pub fn new(fd: c_int) -> Result<Self, DisplayError>

The constructor method default returns the Display's interface from shell.

Important traits for Display
impl Write for Display

pub fn from_winszed(size: Winszed) -> Display

The constructor method default returns the Display's interface from shell.

pub fn get_ss(&self) -> bool

The accessor ss returns the value of 'ss_mod'.

pub fn get_mouse(&self) -> (bool, bool, bool, bool)

The accessor method get_mouse returns the value of mouse_handle'.

pub fn get_window_size(&self) -> &Winszed

The accessor get_window_size returns the window size interface.

pub fn set_window_size(&mut self, size: &Winszed)

The mutator set_window_size replaces the window size.

pub fn get_cursor_coords(&self) -> &Coordinate

The accessor get_cursor_coords returns the value of 'oob', that is the coordinates of the cursor.

pub fn get_newline(&self) -> &Newline

The accessor newlines returns the value of 'newline', that contains all newlines that are now displayed on the screen.

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn into_bytes(&self) -> Vec<c_uchar>

Converts a Vector of Character into a byte vector.

pub fn clear(&mut self) -> Result<size_t>

The method clear purges the screen vector.

pub fn resize(&mut self) -> Result<(), DisplayError>

The method resize updates the size of the output screen.

pub fn resize_with(&mut self, size: &Winszed)

pub fn tricky_resize(&mut self, begin: size_t, end: size_t)

The method tricky_resize updates the size of the output screen.

pub fn goto(&mut self, index: size_t) -> Result<size_t>

The method goto moves the cursor position

pub fn goto_home(&mut self) -> Result<size_t>

The method goto_home moves the cursor to the top left of the output screen.

pub fn goto_up(&mut self, mv: size_t) -> Result<size_t>

The method goto_up moves the cursor up.

pub fn goto_down(&mut self, mv: size_t) -> Result<size_t>

The method goto_down moves the cursor down.

pub fn goto_right(&mut self, mv: size_t) -> Result<size_t>

The method goto_right moves the cursor to its right.

pub fn goto_left(&mut self, mv: size_t) -> Result<size_t>

pub fn goto_begin_row(&mut self)

The method goto_begin_row moves the cursor to the beginning of the row

pub fn goto_end_row(&mut self)

The method goto_end_row moves the cursor to the end of the row

pub fn goto_coord(&mut self, current: Coordinate)

The method goto_coord moves the cursor to the given coordinates

pub fn scroll_down(&mut self, base: size_t)

The method scroll_down append an empty line on bottom of the screen (the cursor doesn't move)

pub fn scroll_up(&mut self, base: size_t)

The method scroll_up insert an empty line on top of the screen (the cursor doesn't move)

pub fn save_position(&mut self)

The method save_position save a position in the variable 'save_position' to get restored with self.table.restore_position() described right after. If save_position() is called many times, only the newest safe will be kept.

pub fn restore_position(&mut self)

The method restore_position move the cursor to coordinates safe with self.table.save_position() described right before. If no coordinates were safe, cursor moves to the top left of the output screen

pub fn insert_empty_line(&mut self, mv: size_t)

The method insert_empty_line insert an empty line on the right of the cursor (the cursor doesn't move)

pub fn erase_right_line(&mut self, current: size_t)

The method erase_right_line erase the current line from the cursor to the next '\n' encountered (char under the cursor included)

pub fn erase_left_line(&mut self, current: size_t)

The method erase_left_line erase the current line from the previous '\n' to the cursor (char under the cursor included)

pub fn erase_line(&mut self, mv: size_t)

The method erase_line erase the entire current line

pub fn erase_up(&mut self)

The method erase_up erase all lines from the current line up to the top of the screen, and erase the current line from the left border column to the cursor. (char under the cursor included)

pub fn erase_down(&mut self)

The method erase_down erase all lines from the current line down to the bottom of the screen and erase the current line from the cursor to the right border column (char under the cursor included)

pub fn print_enter(&mut self)

The method print_enter reproduce the behavior of a '\n'

pub fn print_char(&mut self, first: char, next: &[u8]) -> Result<usize>

The method print_char print an unicode character (1 to 4 chars range)

pub fn catch_numbers<'a>(
&self,
acc: Vec<size_t>,
buf: &'a [u8]
) -> (Vec<size_t>, &'a [u8])

pub fn next_tab(&self) -> size_t

The method next_tab return the size of the current printed tabulation

pub fn erase_chars(&mut self, mv: size_t)

The method erase_chars erases couple of chars in the current line from the cursor.

pub fn insert_chars(&mut self, mv: size_t)

The method erase_chars erases couple of chars in the current line from the cursor.